Small fixes as well as simplifying Cookie list structure
Global_State uses a linked-list of cookies. This patch makes the
linked-list add new cookies to the beginning instead of the end (Getting
rid of the need to have a pointer to the end of the cookie list, and
simplifying some code.)

This patch also has some minor syntax fixes.