Basic UI updates and Reddit API improvements
The UI now detects the terminal size and adjusts output acordingly.
Each link is now limited to one line of text on the screen and
scrolling through the links works correctly.

The Reddit API now implements all the links inside of a RedditLinkList
structure as a simple array of pointers instead of a linked-list
structure (Since random-access to individual links is usually needed, a
linked-list is not very efficent and shouldn't have been used to begin
with).