Implemented viewable comments in main.c
Displaying comments was disabled in main.c while libreddit was being
worked on. This commit fixes this by implementing a new screen for
comments to be displayed with. By default, comments are viewed like the
links screen, with the difference being that if you press enter the
screen will split with the comments full text on the bottom of the
screen.