Adding new options to Link screen and cleaning up project
While on the Link screen you can new update the list with 'u' and get
more Links for the list with 'L'.

'Free' functions in libreddit now simply return if you send them a NULL
pointer.

The 'erase()' call in the drawing functions was removed, and instead
blank lines are printed in the areas that need to be cleared. The
advantage here is that ncurses will optomise which parts of the screen
to redraw if you don't call erase() (Which forces it to redraw
everything). The result is much lower CPU usage and a faster program.