libreddit: Strip unneeded symbols
This patch adds functionality to the Makefile to use objcopy to restrict
the symbols inside of libreddit.a. Internally, functions which are used
across files in libreddit are defined as global. That can create
unexpected symbol conflicts. This patch uses objcopy with a list of
symbols to leave global, and objcopy takes care of localizing the rest
of the functions.