Change libreddit to be compiled as a shared-object
This patch modifies the Makefile structure to compile and use libreddit
as a shared-object by default. The big advantage is in the programming
of libreddit, which can use the visibility attribute to have better
control over what symbols are exported. Also an advantage is this makes
it easier to use libreddit separately.

Because of this change, 'make install' now installs the libreddit
library into PREFIX/lib and reddit.h into PREFIX/include, as well as
creddit into PREFIX/bin. Without installing libreddit, the OS won't be
able to find the .so file when it runs creddit. Because of this, the
option to compile libreddit statically is still an option. Compiling
with:

make STATIC=y

Will compile libreddit as a static library and link creddit with it,
which coresponds roughly to the old behavior.