creddit: Link against libncursesw
This small patch changes the project to link against libncursesw instead
of standard libncurses. When the project moved to using wchar_t instead
of standard char and started using the ncurses wchar_t functions, the
linker line in the Makefile wasn't changed. Some systems will link
agains the wchar_t version of ncurses automatically, but some require it
to be specified and will have linker problems without this change.