creddit: Add code to Makefile setup to check the running system
This commit adds the file common.mk, which runs various commands on the
target system and sets macros that can be used in other Makefiles to
compile with different settings on different machines.

Currently, the only setting is 'umake -r', which checks the running
Kernel (Currently just checked for 'Linux' or 'Drawin'). In creddit.mk,
this is used to decide which ncurses library to link against.