libreddit: Refactor EscParsing code to make it more manageable/useful
This patch does a few things. The first is refactor the code in the
redditParseEscCodes functions into a single function called by the two
(So, one function which handles both char* and wchar_t*).

Secondly, it changes the parameters used on the redditParseEscCodes to
simply take the text and the length of the text (The length parameter is
so you can avoid calculating it if you already have).

Third, it makes the redditParseEscCodes functions exported outside of
the library. This adds some extra functionality to libreddit, and
because of this feature, struct's now include an un-parsed version of
text as well as the parsed version, and the parsed wchar_t version.
While this is a bit more memory consuming, I don't see it being a huge
issue.