libreddit: Integrate EscCode parsing into the Token Parser
This patch implements functionality into the token parser to be able to
handle escaped text, call the EscCode parser, and assign the outputs to
the correct variables.

This is implemented because without it, EscCode parsing requires
somewhat long and tedious callbacks to achieve a good functionality.
This makes the overall code easier to handle and a bit smaller.