creddit: Fix text being left on screen after a Link or comment is closed
There is a bug in the current code where if you close a Link or Comment,
and the Link/Comment was being displayed on lines past the end of the
normal Link/Comment list, then it's left on the screen until you scroll
the list enough to over-write it. This patch shifts some code around
(And adds a little code) so that the space below the Link/Comment list
is cleared via writing spaces over it before the screen is updated.