creddit: libreddit: Add programwide debugging options
This patch adds a few helpful things in the way of debugging. With this
patch, if 'REDDIT_DEBUG' is set to 'y', then the default CFLAGS will not
have -O2 and instead of -g and -DREDDIT_DEBUG (-g compiles with extra
debugging information. -DREDDIT_DEBUG just sets a macro with the name
'REDDIT_DEBUG').

Also added is a file called 'run.sh', which is a short bash script
which runs creddit using LD_PRELOAD to use the version of libreddit.so
inside of the ./bulid folder. If you use ./run.sh to test, then you
won't need to do a make install every time you recompile libreddit.so.
It also will run creddit inside of valgrind if you have REDDIT_DEBUG set
to 'y'.