libreddit: creddit: Implement the 'morechildren' API call
The 'morechildren' API call is used by reddit to gather hidden replies
to a comment. Replies are hidden simply when Reddit's algorithms decide
it's not worth displaying by default, and the libreddit API function
'redditGetCommentChildren'.

This patch also implements the use of the morechildren call into
cReddit's comment displaying screen. You can now see the number of
hidden replies to a comment, and pressing 'm' while a comment with
hidden replies is selected will trigger the morechildren call and add
the comments to the displayed list. (Currently, this is done by
rerendering the entire list. In the future it could probably be done a
bit more 'elegantly').