libreddit: Fix off-by-one error in morechildren call
This patch fixes an off-by-one error in the morechildren call. The error
itself was caused by not subtacting one from the number telling us the
size of an array. The result is that the program tries to call free on a
location past the end of the array, which may end-up causing a
seg-fault.