creddit: libreddit: Project-wide debug option
This patch expands the debug system to be a project-wide debug setup
instead of internal to each module. The basic idea is that
./include/debug.h is included by every module in every file and has
basic macros for DEBUG_PRINT. Each module takes the liberty of setting a
module name for itself for the debug log, as well as the FILE* to write
to. The actual used FILE* can be found in ./src/global.c, all the other
FILE* for the debug file in other modules gets set to the contents of
that one when the program starts and the log file is opened.